requirements.txt
Django==2.2.3
django-datatable-view==0.9.0
django-jquery==3.1.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
